LayoutTests/js/regexp-range-out-of-order-expected.txt

Test for bug 16129: REGRESSION (r27761-r27811): malloc error while visiting http://mysit.es (crashes release build).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS /^[s{-.[]()]$/ threw exception SyntaxError: Invalid regular expression: range out of order in character class.
PASS successfullyParsed is true

TEST COMPLETE
